requirements.txt
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
beir==2.0.0
blis==0.7.11
build==1.0.3
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
confection==0.1.4
contourpy==1.2.0
crcmod==1.7
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
datasets==2.17.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
effdet==0.4.1
einops==0.7.0
elastic-transport==8.12.0
elasticsearch==8.12.1
elasticsearch-dsl==8.12.0
emoji==2.10.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
evaluate==0.4.1
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.110.0
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
gast==0.5.4
google-auth==2.28.1
googleapis-common-protos==1.62.0
grpcio==1.62.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.1.2
iopath==0.1.10
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.9
langchain-community==0.0.25
langchain-core==0.1.28
langchain-elasticsearch==0.1.0
langchain-experimental==0.0.53
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langchainhub==0.1.14
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.1.10
layoutparser==0.3.4
lxml==4.9.4
Markdown==3.5.2
MarkupSafe==2.1.5
marshmallow==3.21.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
mmh3==4.1.0
modelscope==1.12.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
open-clip-torch==2.24.0
openai==1.12.0
opencv-python==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
oss2==2.18.4
overrides==7.7.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pexpect==4.9.0
pikepdf==8.13.0
pillow==10.2.0
pillow_heif==0.15.0
platformdirs==4.2.0
portalocker==2.8.2
posthog==3.4.2
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.25.3
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
pypandoc==1.13
pyparsing==3.1.1
pypdf==4.0.2
pypdfium2==4.27.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pytrec-eval==0.5
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
sentencepiece==0.2.0
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.27
srsly==2.4.8
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
thinc==8.2.3
threadpoolctl==3.3.0
tiktoken==0.6.0
timm==0.9.16
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.1
typer==0.9.0
types-requests==2.31.0.20240218
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
unstructured==0.12.5
unstructured-client==0.18.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.3.4
websocket-client==1.7.0
websockets==12.0
wikipedia==1.4.0
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
